refactor: move timezone helpers to utils
